+++ hubdle-gradle-plugin/main/kotlin/com/javiersc/hubdle/extensions/kotlin/multiplatform/targets/KotlinMultiplatformJvmAndAndroidExtension.kt
@@ -11,12 +11,7 @@ public open class KotlinMultiplatformJvmAndAndroidExtension :
 
     override var Project.isEnabled: Boolean
         get() = hubdleState.kotlin.multiplatform.jvmAndAndroid.isEnabled
-        set(value) =
-            hubdleState.kotlin.multiplatform.run {
-                android.isEnabled = value
-                jvm.isEnabled = value
-                jvmAndAndroid.isEnabled = value
-            }
+        set(value) = hubdleState.kotlin.multiplatform.run { jvmAndAndroid.isEnabled = value }
 
     public override val name: String = "jvmAndAndroid"
 }