+++ hubdle-gradle-plugin/main/kotlin/com/javiersc/hubdle/extensions/kotlin/multiplatform/targets/KotlinMultiplatformMinGWExtension.kt
@@ -11,13 +11,7 @@ public open class KotlinMultiplatformMinGWExtension :
 
     override var Project.isEnabled: Boolean
         get() = hubdleState.kotlin.multiplatform.mingw.isEnabled
-        set(value) =
-            hubdleState.kotlin.multiplatform.run {
-                mingw.isEnabled = value
-                mingwX64.isEnabled = value
-                mingwX86.isEnabled = value
-                native.isEnabled = value
-            }
+        set(value) = hubdleState.kotlin.multiplatform.run { mingw.isEnabled = value }
 
     override val name: String = "mingw"
 }