+++ hubdle-gradle-plugin/main/kotlin/com/javiersc/hubdle/extensions/kotlin/multiplatform/targets/KotlinMultiplatformTvOSExtension.kt
@@ -11,14 +11,7 @@ public open class KotlinMultiplatformTvOSExtension :
 
     override var Project.isEnabled: Boolean
         get() = hubdleState.kotlin.multiplatform.tvos.isEnabled
-        set(value) =
-            hubdleState.kotlin.multiplatform.run {
-                tvos.isEnabled = value
-                tvosArm64.isEnabled = value
-                tvosSimulatorArm64.isEnabled = value
-                tvosX64.isEnabled = value
-                darwin.isEnabled = value
-            }
+        set(value) = hubdleState.kotlin.multiplatform.run { tvos.isEnabled = value }
 
     override val name: String = "tvos"
 }