+++ hubdle-gradle-plugin/main/kotlin/com/javiersc/hubdle/extensions/kotlin/multiplatform/targets/KotlinMultiplatformWAsmExtension.kt
@@ -11,12 +11,7 @@ public open class KotlinMultiplatformWAsmExtension :
 
     override var Project.isEnabled: Boolean
         get() = hubdleState.kotlin.multiplatform.wasm.isEnabled
-        set(value) =
-            hubdleState.kotlin.multiplatform.run {
-                wasm.isEnabled = value
-                wasm32.isEnabled = value
-                native.isEnabled = value
-            }
+        set(value) = hubdleState.kotlin.multiplatform.run { wasm.isEnabled = value }
 
     override val name: String = "wasm"
 }