+++ hubdle-gradle-plugin/main/kotlin/com/javiersc/hubdle/extensions/kotlin/multiplatform/targets/KotlinMultiplatformWatchOSExtension.kt
@@ -11,16 +11,7 @@ public open class KotlinMultiplatformWatchOSExtension :
 
     override var Project.isEnabled: Boolean
         get() = hubdleState.kotlin.multiplatform.watchos.isEnabled
-        set(value) =
-            hubdleState.kotlin.multiplatform.run {
-                watchos.isEnabled = value
-                watchosArm32.isEnabled = value
-                watchosArm64.isEnabled = value
-                watchosSimulatorArm64.isEnabled = value
-                watchosX64.isEnabled = value
-                watchosX86.isEnabled = value
-                darwin.isEnabled = value
-            }
+        set(value) = hubdleState.kotlin.multiplatform.run { watchos.isEnabled = value }
 
     override val name: String = "watchos"
 }